More errors and error checks
Yay, getting ready for the first release. This means that the program is
tested on different operating systems having Steam installed in different
locations (or having it not installed at all). This is to prevent
questions like "hurrdurr, I don't have Steam installed, why can't I
install Skins for it?".
Also, add trailing . to every exception message so it looks more user
friendly.